refactor and addressed some commnet on the PR
