+++ app/forms/amends/AmendExportedByAnotherBusinessFormProvider.scala
@@ -14,20 +14,21 @@
  * limitations under the License.
  */
 
-package forms
+package forms.amends
 
 import forms.mappings.Mappings
-import javax.inject.Inject
 import play.api.data.Form
 
-class AmendExportedWeightFormProvider @Inject() extends Mappings {
+import javax.inject.Inject
+
+class AmendExportedByAnotherBusinessFormProvider @Inject() extends Mappings {
 
   def apply(): Form[Long] =
     Form(
       "value" -> long(
-        "amendExportedWeight.error.required",
-        "amendExportedWeight.error.wholeNumber",
-        "amendExportedWeight.error.nonNumeric")
-          .verifying(inRange(0L, 99999999999L, "amendExportedWeight.error.outOfRange"))
+        "amendExportedByAnotherBusiness.error.required",
+        "amendExportedByAnotherBusiness.error.wholeNumber",
+        "amendExportedByAnotherBusiness.error.nonNumeric")
+          .verifying(inRange(0L, 99999999999L, "amendExportedByAnotherBusiness.error.outOfRange"))
     )
 }
