+++ app/forms/amends/AmendExportedWeightFormProvider.scala
@@ -14,20 +14,21 @@
  * limitations under the License.
  */
 
-package forms
+package forms.amends
 
 import forms.mappings.Mappings
-import javax.inject.Inject
 import play.api.data.Form
 
-class AnotherBusinessExportWeightFormProvider @Inject() extends Mappings {
+import javax.inject.Inject
+
+class AmendExportedWeightFormProvider @Inject() extends Mappings {
 
   def apply(): Form[Long] =
     Form(
       "value" -> long(
-        "anotherBusinessExportWeight.error.required",
-        "anotherBusinessExportWeight.error.wholeNumber",
-        "anotherBusinessExportWeight.error.nonNumeric")
-          .verifying(inRange(0L, 99999999999L, "anotherBusinessExportWeight.error.outOfRange"))
+        "amendExportedWeight.error.required",
+        "amendExportedWeight.error.wholeNumber",
+        "amendExportedWeight.error.nonNumeric")
+          .verifying(inRange(0L, 99999999999L, "amendExportedWeight.error.outOfRange"))
     )
 }
