+++ conf/app.routes
@@ -191,12 +191,12 @@ POST       /change-another-business-export-weight                        control
 
 GET        /amending-exported-plastic-packaging                       controllers.amends.AmendExportedPlasticPackagingController.onPageLoad()
 
-GET        /amend-exported-weight                  controllers.AmendExportedWeightController.onPageLoad(mode: Mode = NormalMode)
-POST       /amend-exported-weight                  controllers.AmendExportedWeightController.onSubmit(mode: Mode = NormalMode)
-GET        /change-amend-exported-weight                        controllers.AmendExportedWeightController.onPageLoad(mode: Mode = CheckMode)
-POST       /change-amend-exported-weight                        controllers.AmendExportedWeightController.onSubmit(mode: Mode = CheckMode)
+GET        /amend-exported-weight                  controllers.amends.AmendExportedWeightController.onPageLoad(mode: Mode = NormalMode)
+POST       /amend-exported-weight                  controllers.amends.AmendExportedWeightController.onSubmit(mode: Mode = NormalMode)
+GET        /change-amend-exported-weight                        controllers.amends.AmendExportedWeightController.onPageLoad(mode: Mode = CheckMode)
+POST       /change-amend-exported-weight                        controllers.amends.AmendExportedWeightController.onSubmit(mode: Mode = CheckMode)
 
-GET        /amend-exported-by-another-business                  controllers.AmendExportedByAnotherBusinessController.onPageLoad(mode: Mode = NormalMode)
-POST       /amend-exported-by-another-business                  controllers.AmendExportedByAnotherBusinessController.onSubmit(mode: Mode = NormalMode)
-GET        /change-amend-exported-by-another-business                        controllers.AmendExportedByAnotherBusinessController.onPageLoad(mode: Mode = CheckMode)
-POST       /change-amend-exported-by-another-business                        controllers.AmendExportedByAnotherBusinessController.onSubmit(mode: Mode = CheckMode)
+GET        /amend-exported-by-another-business                  controllers.amends.AmendExportedByAnotherBusinessController.onPageLoad(mode: Mode = NormalMode)
+POST       /amend-exported-by-another-business                  controllers.amends.AmendExportedByAnotherBusinessController.onSubmit(mode: Mode = NormalMode)
+GET        /change-amend-exported-by-another-business                        controllers.amends.AmendExportedByAnotherBusinessController.onPageLoad(mode: Mode = CheckMode)
+POST       /change-amend-exported-by-another-business                        controllers.amends.AmendExportedByAnotherBusinessController.onSubmit(mode: Mode = CheckMode)
