+++ test/pages/amends/AmendExportedByAnotherBusinessPageSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package pages
+package pages.amends
 
 import pages.behaviours.PageBehaviours
